Support all length memberExpresions, Remove comments, remove spread tests
